wip: moving EE-specific tests to enterprise dir
