+++ frontend/test/__support__/server-mocks/performance.ts
@@ -1,6 +1,6 @@
 import fetchMock from "fetch-mock";
 
-import type { Config as CacheConfig } from "metabase/admin/performance/types";
+import type { Config as CacheConfig } from "metabase/admin/performance/strategies";
 export function setupPerformanceEndpoints(cacheConfigs: CacheConfig[]) {
   fetchMock.get(
     {
